cAER: fix test utilities compilation by linking with libcaer. Fix double-promotion warning in SSHS_helper.
git-svn-id: https://svn.code.sf.net/p/jaer/code/cAER/trunk@7357 b7f4320f-462c-0410-a916-d9f35bb82d52

+++ ext/sshs/sshs_helper.c
@@ -123,7 +123,7 @@ char *sshsHelperValueToStringConverter(enum sshs_node_attr_value_type type, unio
 			break;
 
 		case FLOAT:
-			sshsHelperAllocSprintf(&valueString, "%g", value.ffloat);
+			sshsHelperAllocSprintf(&valueString, "%g", (double) value.ffloat);
 			break;
 
 		case DOUBLE:
